decoupled coinbase from generate page
